--- src/main/java/kr/ai/nemo/groupparticipants/dto/GroupParticipantsListResponse.java
@@ -1,11 +1,0 @@
-package kr.ai.nemo.groupparticipants.dto;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GroupParticipantsListResponse {
-  private List<GroupParticipantDto> participants;
-}
